pyproject.toml
[tool.poetry]
name = "auth_module"
version = "0.1.0"
description = ""
authors = ["Alemkhan <alemkhan.yergaliev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
pydantic = {extras = ["email"], version = "^1.9.1"}
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
asyncpg = "^0.25.0"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
httpx = "^0.22.0"
orjson = "^3.6.8"
PyJWT = "^2.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
mypy = "^0.950"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
sqlalchemy2-stubs = "^0.0.2-alpha.22"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy",
]
exclude = "(alembic|tests|src/core/decorators)"
mypy_path = "src"

follow_imports = "silent"
ignore_missing_imports = true

check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_any_generics = true
no_implicit_optional = true

warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

strict_optional = true
strict_equality = true


[tool.isort]
profile = "black"
